DiffusionModel/requirements.txt
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.5.0
pytorch_lightning==2.0.0
scipy==1.10.1
torch==2.0.0
torchvision==0.15.1
